requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-Login==0.5.0
ebooklib==0.17.1
PyPDF2==1.26.0
Pillow==9.5.0
python-dotenv==0.19.0
Werkzeug==2.0.1
SQLAlchemy==1.4.46